StockTrader/requirements.txt
kiteconnect==4.2.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
requests>=2.31.0
python-dotenv>=0.10.0